# api/openapi.yaml
openapi: 3.1.0
info:
  title: Flex API
  version: 0.5.0
  description: API du serveur multimédia personnel Flex.
servers:
  - url: /api
security:
  - cookieAuth: []
paths:
...
  /media:
    get:
      operationId: listMedia
      summary: Liste les vidéos indexées d'une bibliothèque.
      parameters:
        - $ref: "#/components/parameters/LibraryIdQuery"
      responses:
        "200":
          description: Liste des vidéos.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/MediaFile"
...
  /media/{mediaId}/stream:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    get:
      operationId: streamMedia
      summary: Lit directement le fichier vidéo avec prise en charge HTTP Range.
      responses:
        "200":
          description: Fichier vidéo complet.
        "206":
          description: Portion du fichier demandée par le lecteur.
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: flex_session
  parameters:
...
    LibraryIdQuery:
      name: libraryId
      in: query
      required: true
      schema:
        type: string
    MediaId:
      name: mediaId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    MediaFile:
      type: object
      required: [id, libraryId, filename, sizeBytes, durationMs, width, height, container, videoCodec, audioCodec, modifiedAt, progressMs, completed, title, description, recordedAt, favorite]
      properties:
        id:
          type: string
        libraryId:
          type: string
        filename:
          type: string
        sizeBytes:
          type: integer
          format: int64
        durationMs:
          type: integer
          format: int64
        width:
          type: integer
        height:
          type: integer
        container:
          type: string
        videoCodec:
          type: string
        audioCodec:
          type: string
        modifiedAt:
          type: string
          format: date-time
        progressMs:
          type: integer
          format: int64
          minimum: 0
        completed:
          type: boolean
        title:
          type: string
        description:
          type: string
        recordedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        favorite:
          type: boolean
...